I tried to push more of the formatting code into P::C::Violation. Now, P::C::Violation will handle converting numeric and literal formats for you.  That way, you can specify a numeric or literal format anywhere that you see the -verbose option.  At this point it probably make sense to call it "-format" instead of "-verbose" anyway.
In the long run, I'm not sure if this was a good idea or not.
Eventually, I think there might be several P::C::Report classes
that display Violations in ASCII or HTML or XML or whatever.  So
perhaps the -verbose and -format features would be dispatched to
those classes instead.