Added framework for declaring policy parameters so Perl::Critic can verify that the profile contains semi-valid.  This is accomplished by the policy_parameters() method, which tells Perl::Critic the names of the parameters that the policy will accept.
However, this does not actually verify that the values of the
parameters are valid.  To do that, the Policy should probably provide
a callback that Perl::Critic can invoke with the parameters before
passing them to the Policy constructor.

I'm not sure if this should go into the 1.0 release.  It depends on
the feedback from the rest of the team.