Modified is_method_call() so it can be used to detect barewords on either side of the dereference operator.  For example, suppose you have:
  Foo::Bar->baz();

So now is_method_call will return true when given either the
"Foo::Bar" element or the "baz" element.  This relies on the
assumption that whenever you have barewords on either side of "->", it
implies that some method is being called on some class or object
reference.  I don't yet know if that is a realistic assumption.

The motivation for doing this is to enable help the
RequireExplicitInclusion policy to detect which packages are used by
the code.