Decided that my last modification to is_method_call was a bad idea. So I reverted is_method_call back to its old behavior.
But now there is a new method, is_class_name(), which returns true
when given a bareword that is on the left side of "->".  For example,
it would return true for the "Foo::Bar" element in these expressions:

  my $whatever = Foo::Bar->new()
  Foo::Bar->quux();

The terminology used in the code is starting to get a little confusing.
Words like "class", "package", "method", "subroutine" are starting to
get jumbled up.  Also don't have a good convention of knowing when an
variable refers to a real PPI element, or just the string-ified form
of the element.  Will have to sort this out later.