Holy smokes, it works!
Using the --warn-about-useless-no-critic option, I found several places
where we were disabling too many policies, or were disabling them
due to bugs that have since been fixed.  This is a pretty cool feature,
but I have reservations about making it into a regular Policy.  Since
we use pattern-matching to select the Policies, it is really easy for
someone to disable a *future* policy.  So I'm not sure I want to force
them to change their '## no critic' markers just because they've upgraded 
P::C.

BTW, I discovered that we were using the wrong syntax in some of our
'## no critic' markers:

  ## no critic   'SomePolicy'  # wrong, disables all policies.
  ## no critic   (SomePolicy)  # ok
  ## no critic qw(SomePolicy)  # ok

We should probably emit a warning or something if we come across
a '## no critic' marker that looks wrong.