Merge r3240 to the backlog branch.
Proposed implementation of 'Enhance statistics' TODO.pod item. The
'Data section count' in the TODO is interpreted to mean number of
lines in the data section, not number of data sections, which I
understand to be either 0 or 1. For my own pragmatic reasons, POD in a
data section is counted as POD, not data. The
lib/Perl/Critic/Statistics.pm code now contains a '## no critic
(ProhibitCascadingIfElse)'. If anyone knows a better way to code this,
feel free. Or tell me what is wanted. I have not yet removed the
relevant TODO.pod item, but have commented it.