Merge r3242 to the backlog branch.
RT 43910: is( is_foo(), 1); dangerous. The code in the RT report
appears to be from t/05_utils.t. A sweep with a custom policy failed
to find any examples of this anywhere else. Is there any place to put
a policy whose sole use is to look for this kind of thing in
Perl::Critic tests? If so, would such a test be desirable?