Merge r3279 to the backlog branch.
RT 44515: ProhibitMutatingListFunctions should not prohibit
nonmutating tr///. The change allows tr/fubar// (no substitute string)
with no flag or with /c, but not with /d or /s. It also allows
tr/fubar/fubar/ (match string equals substitute string) with no flag
or with /d, but not /c or /s. Testing with Readonly::Scalar says that
the allowed cases truly do not alter the operand, as opposed to
changing it into itself.