Merge r3286 and r3287 to the backlog branch.
Clean up code before messing it up again.

38942: RegularExpressions::ProhibitUnusedCapture false positive. The
commit 3267 logic that checks for the regex being in the condition of
a while() loop, and if so prevents the extension of the capture
variable list, was extended to include the while() loop's block as
well. The TODO test was made no longer TODO.