Merge most of r3563 to the backlog branch.
Policy Modules::RequireFilenameMatchesPackage now respects the logical
filename as defined by the prevailing #line directive at the point
where the package declaration is found.  Test cases are included.

The r3563 includes moving the $LOCATION_* constants out of Violation
into Utils::Constants.  This part of the commit was not merged because
those constants should not be exposed outside of Violation.  There are
proper accessors to get at the values.  If I could, I'd get rid of the
location() method entirely.