Merge r363[89] to the backlog branch.
Turned PolicySummary.pod.PL into a "policysummary" action in
P::C::Module::Build.  The intent is move the generation of
PolicySummary.pod from a build-time event to a distro-creation-time
event.  That way, the file will be present in the distribution and
available on http://search.cpan.org.

Changed the POD in P::C::Module::Build so that it can automatically
generate online documentation for each of the custom actions.  For
example, when you say: "./Build help authortest".

If the actions are not listed as =items within the ACTION section of
the POD, then it doesn't work.