Further tweak docs.
My read on pod2html is that it does _not_ recognize C<< L<Foo> >> as
containing a link, but neither does it format it the way one (at least
this one) would like.

And who knows what goes on in the minds and hearts of other (or older)
POD translaters? The Shadow may, but I haven't a clue.