RT 52029: state should be allowed by Perl::Critic::Policy::CodeLayout::ProhibitParensWithBuiltins
The minimal implementation of adding 'state' to @ALLOWED seemed the
best. If 'state' is not a builtin, by definition it is not subject to
this policy. If it is a builtin, Perl requires the parens if multiple
variables are specified, same as for 'my', so the parens should not be a
violation.

Added a test for this to t/CodeLayout/ProhibitParensWithBuiltins.run.