RT 43291: t/20_policy_pod_spelling - wrong number of tests for Non English - patch included RT 48986: t/20_policy_pod_spelling.t test failure in non-English locale
Both tickets referred to the same problem -- the wrong number of tests
being skipped in a non-English locale. Both included similar patches,
which created a variable to hold the number of tests, and then used this
in both 'plan ( tests => ... )' and 'skip ...'. The committed code is
similar to the patches, but defines the variable as Readonly::Scalar.

The committed code also includes a fairly ugly change to cover the case
when aspell dies for whatever reason (e.g. not finding a dictionary for
the current locale.) It was tested under Fedora 12 while logged in as
Español de España. Some Documentation::PodSpelling failures were
uncovered once I got aspell working right in English; those were not
fixed in this commit.