RT 55600: ::RequireConstantVersion ($foo = $VERSION) =~ s/// case creates false positive
The title is a bit of an oversimplification; the actual example
submitted is

(my $BAR = sprintf q{%s/%s}, __PACKAGE__, $VERSION) =~ s{o\z}{}xms;

It looked to me like the thing that sanitized the $VERSION was not the
assignment (though that would do it too)  but the fact that it went
through sprintf. So I generalized, and rewrote my original mess into a
new, improved (?) mess. This incorporates two ideas:

1) If the $VERSION is preceded by any operator that produces a new value
(i.e. most except for the boolean logicals), it's OK. The previous code
just checked for '='.

2) If the $VERSION appears to be an argument to a subroutine, it's OK.
You have to skip over subroutines whose argument lists are siblings to
the $VERSION.

I'm sure we can still come up with scenarios that produce false
positives, but I confess to being, at this point, not smart enough to
write a foolproof does_this_variable_get_modified().

The submitted false positive, and a selection of similar code, has been
added to the .run file.