RT #56627: Too many arguments (sub prototype, false positive)
If the subroutine had a prototype, the previous code computed the number
of arguments as the number of characters in the prototype, less two for
the parens. The submitted case had embedded whitespace, which is not
usual, but other characters (semicolons, backslashes, square brackets)
also trigger the problem. The fix was to use tr/// to count the
characters that actually represent arguments.