RT #56619 redux
After discussion on the mailing list, added option
allow_in_main_unless_in_subroutine to
P::C::P::ErrorHandling::RequireCarping to control whether 'warn' or
'die' are allowed unconditionally in main:: (unless in a subroutine).
This option is false by default.

The configuration documentation was updated not only to cover this, but
to try to make explicit its interaction with
allow_messages_ending_in_newlines (it's ||, not &&).

The original work was committed as revision 3799.