Tell policy BuiltInFunctions::ProhibitLvalueSubstr not to complain when there is a low-precedence operator between the substr() and the assignment operator.
Note that the modified code assumes that anything with the same
precedence as '=' is an assignment operator. The hash that drives the
precedence_of() subroutine contains a bunch of things at the same
precedence as '='. All appear to be things like '||=', which PPI
decomposes into '||' and '=' anyway. But if someone's religion wants a
test for a literal '=', I will be glad to oblige.