Add new policy InputOutput::RequireEncodingWithUTF8Layer. This complains if it sees ':utf8' rather than ':encoding(utf8)' in open() or binmode(). It is severity 5.
The issue is that ':utf8' flags the string as UTF8 without validating
it. Invalid UTF8 can have bad effects in your code. Exhibit "A" is an
exploit reported on PerlMonks where a script running with -T read input
and sanitized it with m/^(\w+)$/, after which $1 contained shell
meta-characters. This leads to the same kind of problems that the
two-argument open() has, and that policy is severity 5.