RT #70901: Policies missing operators
The issue was missing augmented assignment operators. Patches were
included (and accepted). The only policy patched was
ControlStructures::ProhibitMutatingListFunctions, but '//=' was also
missing from Perl::Critic::Utils::precedence_of().

I'm not sure how serious this was, since PPI parses most of these as two
operators (that is, '+=' gets parsed as '+', '='). But '//=' appears to
be an exception, and is therefore missed both in the policy and anything
that uses precedence_of().