RT #74429: 'Negative array index should be used' a little misleading
The problem was that the documentation of
Variables::RequireNegativeIndices did not mention at all the fact
the semantics of $#foo-$n change when $n becomes greater than $#foo,
whereas the semantics of -$n do not. The author of the ticket seems to
have been burned when trying to straighten out non-compliances.

The long explanation of a non-compliance has been revised to feature the
semantics change prominently.

The short explanation has not been changed, because I confess my
inability to do significantly better without taking significantly more
verbage.