RT #77510: Left curlys as literals in regexps are deprecated.
The use of left curly brackets ("{") as literals in regular expressions
is deprecated, and as of Perl 5.17.0 produces a compiler warning. Given
the code policies required for self-compliance, the solution is to make
them into a singleton character class (i.e. "[{]").

A change in the version of PPIx::Regexp was also done, because that also
turned out to have literal curlies (blush!).