RT #79138: RequireArgUnpacking confused by @_ in finally{}
This policy contains some logic for testing the size of the argument
list -- basically allowing the combinations '$something == @_', '@_ ==
$something', '$something != @_, and '@_ != $something'.

My read on the ticket is that finally{} is not involved, but that the
requestor is asking for the size test logic to be more comprehensive.

This I have done, adding the other comparison operators, plus boolean
operators, plus logic to handle suffix conditionals.

The .run file has an added test, though the test does not cover the
Cartesian product of all the possibilities.