Get rid of test generator for testing without optional deps.
Going forward, there will be no optional deps.  If it makes
the application work better, then it will be a required dep.
We are not afraid of dependencies, so there is no reason to
compromise the user experience for their sake.

Also, I moved the memory leak test into the xt/ area.