Fix test to accommodate changes in Config::Tiny
Starting in version 2.17, Config::Tiny emits a slightly
different exception message when the config file does not
exist.  This caused our test to fail.  So now, we only test
the part of the message that Perl::Critic produces.  The
test isn't quite as good, but it is less brittle.

This fixes #16.