Remove workaround for PPI parsing bug on &&= and ||=
As of version 1.216_01, PPI correctly parses these as
one operator instead of two.  So we need to remove our
workaround and require the newer version of PPI.

I feel icky about listing a dev version as a prereq.
We should update that and cut a new version of P::C
as soon as PPI makes a production release.