Fix the tracking of violation lines:
The code that tries to discover on which line a violation occurred
uses a regex that accidentally treats multiple newlines as a
single line.

Using a policy that tries to detect errors in multiple lines,
such as Perl::Critic::Policy::CodeLayout::ProhibitHashBarewords:

  Use of uninitialized value $min_width in numeric gt (>)
      at /usr/local/share/perl/5.18.2/String/Format.pm line 51.

  Use of uninitialized value $replacement in concatenation (.) or string
      at /usr/local/share/perl/5.18.2/String/Format.pm line 67.

The solution is to remove '\s*' and instead make sure we split
with a negative third argument in order to preserve the number of
consecutive empty lines we encounter.