Add RequireArgUnpacking  option 'allow_closures'
One way to validate subroutine arguments is with a closure. Specio
(recently installed in DateTime) uses this pattern. But there appears to
be no way to tell Subroutines::RequireArgUnpacking to permit this.

This commit adds configuration option 'allow_closures' to permit this.
The option is disabled by default. Tests have been added
t/Subroutines/RequireArgUnpacking.run.