Use Readonly variable, clarify documentation:
Travis CI checks for - and fails over - using "-1" as an arbitrary
value in the split() command. This commit adds a variable with Readonly.

I've also added documentation for this, since I didn't remember exactly
why it has fixed it. (Sure, the commit contains the data, but it's good
to see when inspecting the code. Feel free to remove.)