Replace GPL Version 2 license text with GPL version 1 license text
The Perl license is "GPL (Version 1 or later) or Artistic (Version 1)",
but the LICENSE file includes the text of GPL Version 2. Although the
Version 2 license is a valid one under which the code can be distributed,
it is surely better to include the base license referred to at the top
of the file.

The GPL Version 1 text is taken verbatim from the GNU website at
https://www.gnu.org/licenses/old-licenses/gpl-1.0.txt, with just the
line feed characters and appendix "How to Apply These Terms to Your New
Programs" removed.

This change has the added benefit of fixing the FSF snail mail address
referred to in the LICENSE file, which has been out of date for many years
now.