Fix fencepost error in scope calculation
If the outer element does not define a scope, Perl::Critic::Document's
element_is_in_lexical_scope_after_statement_containing() needs to check
if the inner element is after the outer element. The check was
erroneously succeeding if the inner element was the last element in the
outer element.

The fix was twofold:

1) If the last element in the outer element is a PPI::Node, descend into
its last element, and so on until we have the actual last element in the
statement.

2) Use '>=' rather than '>' in the comparison, since we are comparing
the start of the inner element to the start of the last element in the
outer element.