Recode all_perl_files() to use File::Find
The current implementation of all_perl_files() uses opendir/readdir.
This is usually adwquate, but is badly behaved on distributions that are
pathological enough to include circular symbolic references.

As a side effect, this rewrite also properly excludes Emacs backup files
from the output; the old code only does so if the file is in the
top-level directory being scanned.