Add policy to prohibit using "\shift"
In Perl 5.21.4, a bug was introduced when C<\shift> is used.  This bug remains
in modern Perl.

In short, the bug causes the ref counter for the aliased variable to be
incremented when running the subroutine, but it is not subsequently decremented
after the subroutine returns.  In addition to leaking memory, this issue can
also delay the cleanup of objects until Global Destruction, which can cause
further issues.

For more information, see https://rt.perl.org/Public/Bug/Display.html?id=126676