Use PPIx::Regexp::Util::is_ppi_regexp_element()
in preference to __is_ppi_regexp_element(). There is no functional
difference, but the former is documented as public and supported,
whereas the latter is documented as package-private and unsupported.

Dependencies were updated to reflect the fact that you need
PPIx::Regexp::Util 0.068 to get the public version of the functionality.