GitHub #888 ProhibitUnusedCapture false positive
The issue here is that, in trying to figure out what captures are used,
the code scans forward from the regexp until it hits what it thinks is
end-of-scope, or another regex. But the regex in a split() needs to be
ignored, because its captures (if any) appear in the output of the
split(), not the caller's capture variables.

This commit attempts to implement this exemption. It handles
    split /.../, ...
and
    split( /.../, ... )
but nothing more exotic.