ProhibitStringyEval: allow package declaration when allow_includes=true
The ProhibitStringyEval policy can be relaxed to allow eval strings that
only import a library file, however this mechanism does not allow eval
strings that import a library file in another package's namespace, i.e.:

    eval "package $pkg; use Module; 1;"

Since the NAMESPACE given to "package" must be a bareword, this cannot
be achieved using the block form of eval, as recommended by
ProhibitStringyEval.

When allow_includes is true, permit import-only eval strings to begin
with a (top-level) package declaration.