RequireUse{Strict,Warnings}: handle `package Foo {}` and more lexical scopes
These two were patched to not be tricked by this:

    { use strict; }
    $foo = 1;

Where the strictness would be scoped, so `$foo` ends up not being
covered.

However, the way it was patched was a heuristic: if the `use strict`
wasn't a file-level statement, it was ignored.  Therefore these:

    package Foo;
    {
        use strict;
    }

    package Foo {
        use strict;
    }

    {
        package Foo;
        use strict;
    }

All failed.

This commit changes the heuristic.  It finds all the `strict` (and
`warnings`) statements in the document, and tries finding which lines
they will cover.

The heuristic fails at least in this case:

    {
        use strict;
    } $foo;

which is mis-handled as covered by `strict`, because `$foo` happens
to be on the same line that ends the lexical scope where `strict` was
enabled.