Issue #878 bareword filehandle dies on `open(CHECK, '/foo');` (#899)
* Fix #843 false positive in ProhibitPunctuationVars

Delegated interpretation of regexen to PPIx::Regexp, which is already a
prerequisite.

There are failures in t/05_utils.t and xt/40_criticize-code.t, but they
were present before I started working on this.

* Use PPIx::Regexp::Util::is_ppi_regexp_element()

in preference to __is_ppi_regexp_element(). There is no functional
difference, but the former is documented as public and supported,
whereas the latter is documented as package-private and unsupported.

Dependencies were updated to reflect the fact that you need
PPIx::Regexp::Util 0.068 to get the public version of the functionality.

* on `open(CHECK, '/foo');`

This is the consequence of a mis-parse by PPI.

PPI PR #247 addresses this.

This commit contains a workaround for the PPI bug, which can be removed
if the PR is accepted and published, and defensive code in case
something else unexpected shows up.